dom/indexedDB/test/test_open_for_principal.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
  function* testSteps()
  {
    is("open" in indexedDB, true, "open() defined");
    is("openForPrincipal" in indexedDB, false, "openForPrincipal() not defined");

    is("deleteDatabase" in indexedDB, true, "deleteDatabase() defined");
    is("deleteForPrincipal" in indexedDB, false, "deleteForPrincipal() not defined");

    finishTest();
  }
  </script>
  <script type="text/javascript" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
